Updates after the comments:
Indvidual command tests replaced with generic test cases:
  - ClusterConfigStartStopPauseAndResumeGatewaySenderCommandDUnitTest.java
  - ClusterConfigStartStopPauseAndResumeGatewaySenderOperationDUnitTest
