+++ geode-wan/src/distributedTest/java/org/apache/geode/internal/cache/wan/wancommand/ClusterConfigStartStopPauseAndResumeGatewaySenderOperationDUnitTest.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.cache.wan.serial;
+package org.apache.geode.internal.cache.wan.wancommand;
 
 import static org.apache.geode.distributed.ConfigurationProperties.DISTRIBUTED_SYSTEM_ID;
 import static org.apache.geode.distributed.ConfigurationProperties.REMOTE_LOCATORS;
@@ -34,9 +34,12 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 
+import junitparams.JUnitParamsRunner;
+import junitparams.Parameters;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
 
 import org.apache.geode.cache.Region;
 import org.apache.geode.cache.wan.GatewaySender;
@@ -45,6 +48,11 @@ import org.apache.geode.internal.cache.InternalCache;
 import org.apache.geode.internal.cache.RegionQueue;
 import org.apache.geode.internal.cache.wan.AbstractGatewaySender;
 import org.apache.geode.internal.cache.wan.AbstractGatewaySenderEventProcessor;
+import org.apache.geode.internal.cache.wan.parallel.ConcurrentParallelGatewaySenderEventProcessor;
+import org.apache.geode.internal.cache.wan.parallel.ConcurrentParallelGatewaySenderQueue;
+import org.apache.geode.internal.cache.wan.parallel.ParallelGatewaySenderEventProcessor;
+import org.apache.geode.internal.cache.wan.serial.ConcurrentSerialGatewaySenderEventProcessor;
+import org.apache.geode.internal.cache.wan.serial.SerialGatewaySenderEventProcessor;
 import org.apache.geode.management.internal.cli.util.CommandStringBuilder;
 import org.apache.geode.management.internal.i18n.CliStrings;
 import org.apache.geode.test.dunit.rules.ClientVM;
@@ -54,7 +62,9 @@ import org.apache.geode.test.junit.categories.WanTest;
 import org.apache.geode.test.junit.rules.GfshCommandRule;
 
 @Category({WanTest.class})
-public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Serializable {
+@RunWith(JUnitParamsRunner.class)
+public class ClusterConfigStartStopPauseAndResumeGatewaySenderOperationDUnitTest
+    implements Serializable {
 
   @Rule
   public ClusterStartupRule clusterStartupRule = new ClusterStartupRule(8);
@@ -73,94 +83,92 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
   private ClientVM clientSite2;
 
   /**
-   * Verify that serial gateway-sender state is persisted after pause and resume gateway-sender
+   * Verify that gateway-sender state is persisted after pause and resume gateway-sender
    * commands are executed, and that gateway-sender works as expected after member restart:
    *
-   * 1. Pause gateway-sender
-   * 2. Restart servers that host gateway-sender
-   * 3. Run some traffic and verify that data is enqueued in serial gateway-sender queues
-   * 4. Resume gateway-sender
-   * 5. Verify that traffic is sent over the gateway-sender to remote site
-   * 6. Restart servers that host gateway-sender
-   * 7. Run some traffic and verify that traffic is sent over the gateway-sender to remote site
-   */
-  @Test
-  public void testThatPauseResumeStateRemainAfterTheRestartOfMembers() throws Exception {
-    configureSites(false, "PARTITION", "0");
-
-    executeGfshCommand(CliStrings.PAUSE_GATEWAYSENDER);
-    verifyGatewaySenderState(true, true);
-
-    // stop servers on site #2
-    server1Site2.stop(true);
-    server2Site2.stop(true);
-
-    // start again servers in Site #2
-    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
-    server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
-
-    verifyGatewaySenderState(true, true);
-
-    // Do some puts
-    Set<String> keys = clientSite2.invoke(() -> doPutsInRange(0, 15));
-    // Check that data has been enqueued
-    server1Site2.invoke(() -> checkQueueSize("ln", keys.size()));
-
-    executeGfshCommand(CliStrings.RESUME_GATEWAYSENDER);
-    verifyGatewaySenderState(true, false);
-
-    // Check that data is sent over the gateway-sender
-    clientSite1.invoke(() -> checkDataAvailable(keys));
-
-    // stop servers on site #2
-    server1Site2.stop(true);
-    server2Site2.stop(true);
-
-    // start again servers in Site #2
-    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
-    server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
-    verifyGatewaySenderState(true, false);
-
-    // Do some puts
-    Set<String> key1 = clientSite2.invoke(() -> doPutsInRange(20, 35));
-    // Check that data is sent over the gateway-sender
-    clientSite1.invoke(() -> checkDataAvailable(key1));
-  }
-
-  /**
-   * Verify that gateway-sender queue is persisted while in paused state and it is recovered after
-   * the restart of members.
-   *
-   * - Region type: PARTITION_PERSISTENT and non-redundant
-   * - Gateway sender configured with queue persistence
+   * - Region type: PARTITION and non-redundant
+   * - Gateway sender configured without queue persistence
    *
    * 1. Pause gateway-sender
-   * 2. Run some traffic and verify that data is enqueued in serial gateway-sender queues
-   * 3. Restart servers that host gateway-sender
-   * 4. Check that queues are recovered from persistent storage, and that gateway-sender
-   * remained in paused state
-   * 5. Run some traffic and verify that new data is enqueued in serial gateway-sender queues
-   * 6. Resume gateway-sender
-   * 7. Verify that complete traffic is sent over the gateway-sender to remote site
+   * 2. Run some traffic and verify that data is enqueued in gateway-sender queues
+   * 3. Restart all servers that host gateway-sender
+   * 4. Run some traffic and verify that data is enqueued in gateway-sender queues, old
+   * data should be lost from queue after servers are restarted
+   * 5. Resume gateway-sender
+   * 6. Verify that latest traffic is sent over the gateway-sender to remote site
    */
   @Test
-  public void testThatPauseStateAndQueuesRemainAfterRestart() throws Exception {
-    configureSites(true, "PARTITION_PERSISTENT", "0");
+  @Parameters({"true", "false"})
+  public void testThatPauseStateRemainAfterTheRestartOfMembers(String isParallel) throws Exception {
+    configureSites(false, "PARTITION", "0", isParallel);
 
     executeGfshCommand(CliStrings.PAUSE_GATEWAYSENDER);
     verifyGatewaySenderState(true, true);
     server2Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
     server1Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
 
-    Set<String> keysQueued = clientSite2.invoke(() -> doPutsInRange(70, 85));
-    clientSite2.invoke(() -> checkDataAvailable(keysQueued));
-    server1Site2.invoke(() -> checkQueueSize("ln", keysQueued.size()));
+    // Do some puts and check that data has been enqueued
+    Set<String> keysQueue = clientSite2.invoke(() -> doPutsInRange(0, 15));
+    server1Site2.invoke(() -> checkQueueSize("ln", keysQueue.size()));
 
     // stop servers on site #2
     server1Site2.stop(false);
     server2Site2.stop(false);
 
     // start again servers in Site #2
+    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
+    server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
+
+    verifyGatewaySenderState(true, true);
+
+    // Do some puts and check that data has been enqueued, previous queue should be lost
+    Set<String> keysQueue1 = clientSite2.invoke(() -> doPutsInRange(20, 35));
+    server1Site2.invoke(() -> checkQueueSize("ln", keysQueue1.size()));
+
+    executeGfshCommand(CliStrings.RESUME_GATEWAYSENDER);
+    verifyGatewaySenderState(true, false);
+
+    // check that queue is empty
+    server1Site2.invoke(() -> checkQueueSize("ln", 0));
+    // check that data replicated to remote site
+    clientSite1.invoke(() -> checkDataAvailable(keysQueue1));
+  }
+
+  /**
+   * Verify that gateway-sender queue is persisted while in paused state and it is recovered after
+   * the restart of servers.
+   *
+   * - Region type: PARTITION_PERSISTENT and non-redundant
+   * - Gateway sender configured with queue persistence
+   *
+   * 1. Pause gateway-sender
+   * 2. Restart all servers that host gateway-sender
+   * 3. Run some traffic and verify that data is enqueued in gateway-sender queues
+   * 4. Restart all servers that host gateway-sender
+   * 5. Run some traffic and verify that data is enqueued in gateway-sender queues, old
+   * data should be recovered after servers restarted
+   * 6. Resume gateway-sender
+   * 5. Verify that complete traffic is sent over the gateway-sender to the remote site
+   */
+  @Test
+  @Parameters({"true", "false"})
+  public void testThatPauseStateRemainAfterRestartAllServersPersistent(String isParallel)
+      throws Exception {
+    configureSites(true, "PARTITION_PERSISTENT", "0", isParallel);
+
+    executeGfshCommand(CliStrings.PAUSE_GATEWAYSENDER);
+    verifyGatewaySenderState(true, true);
+    server2Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
+    server1Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
+
+    // Do some puts and check that data has been enqueued
+    Set<String> keysQueue = clientSite2.invoke(() -> doPutsInRange(0, 15));
+    server1Site2.invoke(() -> checkQueueSize("ln", keysQueue.size()));
+
+    // stop servers on site #2
+    server1Site2.stop(false);
+    server2Site2.stop(false);
+
     Thread thread = new Thread(
         () -> server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort()));
     Thread thread1 = new Thread(
@@ -169,178 +177,103 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     thread.start();
     thread1.start();
     thread.join();
-    thread1.join();
+    thread.join();
 
     verifyGatewaySenderState(true, true);
 
-    // queue recovered from persistent storage
-    server1Site2.invoke(() -> checkQueueSize("ln", keysQueued.size()));
-
-    // Do some puts
-    Set<String> keysQueued1 = clientSite2.invoke(() -> doPutsInRange(0, 15));
-    // Check that data has been enqueued
-    server1Site2.invoke(() -> checkQueueSize("ln", keysQueued1.size() + keysQueued.size()));
+    // Do some puts and check that data has been enqueued, previous queue data should not be lost
+    Set<String> keysQueue1 = clientSite2.invoke(() -> doPutsInRange(20, 35));
+    server1Site2.invoke(() -> checkQueueSize("ln", keysQueue1.size() + keysQueue.size()));
 
     executeGfshCommand(CliStrings.RESUME_GATEWAYSENDER);
     verifyGatewaySenderState(true, false);
 
-    // Check that data is sent over the gateway-sender
+    // check that queue is empty
     server1Site2.invoke(() -> checkQueueSize("ln", 0));
-    clientSite1.invoke(() -> checkDataAvailable(keysQueued1));
-    clientSite1.invoke(() -> checkDataAvailable(keysQueued));
+    // check that data replicated to remote site
+    clientSite1.invoke(() -> checkDataAvailable(keysQueue1));
+    clientSite1.invoke(() -> checkDataAvailable(keysQueue));
   }
 
   /**
    * Verify that gateway-sender is recovered from redundant server after the
    * restart of member.
    *
-   * - Region type: PARTITION_PERSISTENT and redundant
-   * - Gateway sender configured with queue persistence
+   * - Region type: PARTITION and redundant
+   * - Gateway sender configured without queue persistence
    *
    * 1. Pause gateway-sender
-   * 2. Run some traffic and verify that data is enqueued in serial gateway-sender queues
-   * 3. Stop one server that host gateway-sender
-   * 4. Resume gateway-sender on running server, and check that data is replicated
-   * 5. Pause gateway-sender again
-   * 5. Run some traffic and verify that new data is enqueued in serial gateway-sender queues
-   * 6. Start server that has been previously stopped, and check that already replicated data is not
-   * inserted into queue
-   * 7. Resume gateway-sender
-   * 8. Verify that complete traffic is sent over the gateway-sender to remote site
+   * 2. Run some traffic and verify that data is enqueued in gateway-sender queues
+   * 3. Restart one server that host gateway-sender
+   * 4. Run some traffic and verify that data is enqueued in gateway-sender queues, old
+   * data should not be lost from queue after servers are restarted because of redundancy
+   * 5. Resume gateway-sender
+   * 6. Verify that queued traffic is sent over the gateway-sender to remote site
    */
   @Test
-  public void testThatPauseStateRemainAfterTheRestartOfMembersRedundantPersistent()
+  @Parameters({"true", "false"})
+  public void testThatPauseStateRemainAfterRestartOneServerRedundant(String isParallel)
       throws Exception {
-    configureSites(true, "PARTITION_PERSISTENT", "1");
+    configureSites(false, "PARTITION", "1", isParallel);
 
     executeGfshCommand(CliStrings.PAUSE_GATEWAYSENDER);
     verifyGatewaySenderState(true, true);
     server2Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
     server1Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
 
-    Set<String> keysQueued = clientSite2.invoke(() -> doPutsInRange(70, 85));
-    server1Site2.invoke(() -> checkQueueSize("ln", keysQueued.size()));
+    // Do some puts and check that data has been enqueued
+    Set<String> keys1 = clientSite2.invoke(() -> doPutsInRange(0, 15));
+    server1Site2.invoke(() -> checkQueueSize("ln", keys1.size()));
 
-    // stop servers on site #2
+    // stop server on site #2
     server1Site2.stop(false);
 
-    executeGfshCommand(CliStrings.RESUME_GATEWAYSENDER);
-    verifyGatewaySenderStateOnMember(server2Site2, true, false);
-    server2Site2.invoke(() -> checkQueueSize("ln", 0));
-    clientSite1.invoke(() -> checkDataAvailable(keysQueued));
+    // start again server in Site #2
+    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
 
-    executeGfshCommand(CliStrings.PAUSE_GATEWAYSENDER);
-    verifyGatewaySenderStateOnMember(server2Site2, true, true);
+    verifyGatewaySenderState(true, true);
 
-    server2Site2.invoke(() -> waitAllDispatcherThreadsToPause("ln"));
-
-    Set<String> keysQueued1 = clientSite2.invoke(() -> doPutsInRange(100, 200));
-    clientSite2.invoke(() -> checkDataAvailable(keysQueued1));
-    server2Site2.invoke(() -> checkQueueSize("ln", keysQueued1.size()));
-
-    // start again servers in Site #2
-    Thread thread = new Thread(
-        () -> server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort()));
-    // start threads
-    thread.start();
-    thread.join();
-
-    verifyGatewaySenderStateOnMember(server1Site2, true, true);
-    server1Site2.invoke(() -> checkQueueSize("ln", keysQueued1.size()));
+    // Do some puts and check that data has been enqueued, previous queue should not be lost
+    // due to configured redundancy
+    Set<String> keys = clientSite2.invoke(() -> doPutsInRange(20, 35));
+    server1Site2.invoke(() -> checkQueueSize("ln", keys.size() + keys1.size()));
 
     executeGfshCommand(CliStrings.RESUME_GATEWAYSENDER);
-    verifyGatewaySenderStateOnMember(server2Site2, true, false);
-    server2Site2.invoke(() -> checkQueueSize("ln", 0));
-    clientSite1.invoke(() -> checkDataAvailable(keysQueued1));
-    clientSite1.invoke(() -> checkDataAvailable(keysQueued));
+    verifyGatewaySenderState(true, false);
+
+    // check that queue is empty
+    server1Site2.invoke(() -> checkQueueSize("ln", 0));
+    // check that data replicated to other site
+    clientSite1.invoke(() -> checkDataAvailable(keys));
+    clientSite1.invoke(() -> checkDataAvailable(keys1));
   }
 
   /**
-   * Verify that serial gateway-sender state is persisted after stop and start gateway-sender
+   * Verify that gateway-sender state is persisted after stop and start gateway-sender
    * commands are executed, and that gateway-sender works as expected after member restart:
    *
-   * 1. Stop gateway-sender
-   * 2. Restart servers that host gateway-sender
-   * 3. Run some traffic and verify that data is just stored in regions, but not enqueued and sent
-   * to the remote site
-   * 4. Start gateway-sender
-   * 5. Verify that new traffic is sent over the gateway-sender to remote site
-   * 6. Restart servers that host gateway-sender
-   * 7. Run some traffic and verify that traffic is sent over the gateway-sender to remote site
-   */
-  @Test
-  public void testThatStopStateRemainAfterTheRestartOfMembers() throws Exception {
-    configureSites(false, "PARTITION", "0");
-
-    executeGfshCommand(CliStrings.STOP_GATEWAYSENDER);
-    verifyGatewaySenderState(false, false);
-
-    // stop servers on site #2
-    server1Site2.stop(true);
-    server2Site2.stop(true);
-
-    // start again servers in Site #2
-    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
-    server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
-
-    verifyGatewaySenderState(false, false);
-
-    // Do some puts
-    Set<String> keys = clientSite2.invoke(() -> doPutsInRange(0, 15));
-    clientSite2.invoke(() -> checkDataAvailable(keys));
-
-    // Check that data is not enqueued
-    server1Site2.invoke(() -> checkQueueSize("ln", 0));
-
-    executeGfshCommand(CliStrings.START_GATEWAYSENDER);
-    verifyGatewaySenderState(true, false);
-
-    // Do some puts
-    Set<String> keys1 = clientSite2.invoke(() -> doPutsInRange(20, 35));
-    clientSite2.invoke(() -> checkDataAvailable(keys1));
-    clientSite1.invoke(() -> checkDataAvailable(keys1));
-
-    // stop servers on site #2
-    server1Site2.stop(true);
-    server2Site2.stop(true);
-
-    // start again servers in Site #2
-    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
-    server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
-    verifyGatewaySenderState(true, false);
-
-    // Do some puts
-    Set<String> key1 = clientSite2.invoke(() -> doPutsInRange(40, 55));
-    // Check that data is sent over the gateway-sender
-    clientSite1.invoke(() -> checkDataAvailable(key1));
-    clientSite1.invoke(() -> checkDataNotAvailable(keys));
-  }
-
-  /**
-   * Verify that gateway-sender queue is not filled in stopped state and it is not recovered
-   * after the restart of member.
-   *
-   * - Region type: PARTITION_PERSISTENT and non-redundant
-   * - Gateway sender configured with queue persistence
+   * - Region type: PARTITION and non-redundant
+   * - Gateway sender configured without queue persistence
    *
    * 1. Stop gateway-sender
-   * 2. Run some traffic and verify that data is not enqueued in serial gateway-sender queues
+   * 2. Run some traffic and verify that data is stored in partition region, and not replicated to
+   * the other site
    * 3. Restart servers that host gateway-sender
-   * 4. Check that queues are recovered empty, and that gateway-sender
-   * remained in stopped state
-   * 5. Run some traffic and verify that data is not enqueued in serial gateway-sender queues
-   * 6. Start gateway-sender
-   * 7. Verify that no traffic is sent over the gateway-sender to remote site
+   * 4. Run some traffic and verify that partition region is recovered and that data is not
+   * replicated to the other site
+   * 5. Start gateway-sender
+   * 6. Run some traffic and verify that traffic is sent over the gateway-sender to remote site
    */
   @Test
-  public void testThatStopStateRemainAfterTheRestartOfMembersPersistent() throws Exception {
-    configureSites(true, "PARTITION_PERSISTENT", "0");
+  @Parameters({"true", "false"})
+  public void testThatStopStateRemainAfterTheRestartOfMembers(String isParallel) throws Exception {
+    configureSites(false, "PARTITION", "0", isParallel);
 
     executeGfshCommand(CliStrings.STOP_GATEWAYSENDER);
     verifyGatewaySenderState(false, false);
 
-    Set<String> keysNotQueued = clientSite2.invoke(() -> doPutsInRange(70, 85));
-    clientSite2.invoke(() -> checkDataAvailable(keysNotQueued));
+    Set<String> keys1 = clientSite2.invoke(() -> doPutsInRange(0, 15));
+    clientSite2.invoke(() -> checkDataAvailable(keys1));
     server1Site2.invoke(() -> checkQueueSize("ln", 0));
 
     // stop servers on site #2
@@ -348,6 +281,57 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     server2Site2.stop(false);
 
     // start again servers in Site #2
+    server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
+    server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
+
+    verifyGatewaySenderState(false, false);
+
+    Set<String> keys = clientSite2.invoke(() -> doPutsInRange(20, 35));
+    clientSite2.invoke(() -> checkDataAvailable(keys));
+
+    executeGfshCommand(CliStrings.START_GATEWAYSENDER);
+    verifyGatewaySenderState(true, false);
+
+    Set<String> keys3 = clientSite2.invoke(() -> doPutsInRange(40, 55));
+    clientSite2.invoke(() -> checkDataAvailable(keys3));
+    clientSite1.invoke(() -> checkDataAvailable(keys3));
+  }
+
+  /**
+   * Verify that colocated partition regions (gws queue and region) are created
+   * after servers are restarted and gateway-sender is created in stopped state.
+   *
+   * - Region type: PARTITION_PERSISTENT and non-redundant
+   * - Gateway sender configured with queue persistence
+   *
+   * 1. Stop gateway-sender
+   * 2. Run some traffic and verify that data is stored in partition region, and not replicated to
+   * the remote site
+   * 3. Restart servers that host gateway-sender
+   * 4. Run some traffic and verify that partition region is recovered and that data is not
+   * replicated to the other site
+   * 5. Start gateway-sender
+   * 6. Run some traffic and verify that onl latest traffic is sent over the gateway-sender
+   * to remote site
+   */
+  @Test
+  @Parameters({"true", "false"})
+  public void testThatStopStateRemainAfterTheRestartOfMembersAndAllRegionsRecover(String isParallel)
+      throws Exception {
+    configureSites(true, "PARTITION_PERSISTENT", "0", isParallel);
+
+    executeGfshCommand(CliStrings.STOP_GATEWAYSENDER);
+    verifyGatewaySenderState(false, false);
+
+    Set<String> keys1 = clientSite2.invoke(() -> doPutsInRange(0, 15));
+    clientSite2.invoke(() -> checkDataAvailable(keys1));
+
+    server1Site2.invoke(() -> checkQueueSize("ln", 0));
+
+    // stop servers on site #2
+    server1Site2.stop(false);
+    server2Site2.stop(false);
+
     Thread thread = new Thread(
         () -> server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort()));
     Thread thread1 = new Thread(
@@ -356,40 +340,46 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     thread.start();
     thread1.start();
     thread.join();
-    thread1.join();
+    thread.join();
 
     verifyGatewaySenderState(false, false);
-    // Do some puts
-    Set<String> keysNotQueued1 = clientSite2.invoke(() -> doPutsInRange(0, 15));
-    // Check that data has been enqueued
-    clientSite2.invoke(() -> checkDataAvailable(keysNotQueued1));
+
+    // check that partition region is created and that accepts new traffic
+    Set<String> keys = clientSite2.invoke(() -> doPutsInRange(20, 35));
+    clientSite2.invoke(() -> checkDataAvailable(keys));
+
     server1Site2.invoke(() -> checkQueueSize("ln", 0));
 
     executeGfshCommand(CliStrings.START_GATEWAYSENDER);
     verifyGatewaySenderState(true, false);
 
-    server1Site2.invoke(() -> checkQueueSize("ln", 0));
-    clientSite1.invoke(() -> checkDataNotAvailable(keysNotQueued));
-    clientSite1.invoke(() -> checkDataNotAvailable(keysNotQueued1));
+    Set<String> keys3 = clientSite2.invoke(() -> doPutsInRange(40, 55));
+    clientSite2.invoke(() -> checkDataAvailable(keys3));
+    clientSite1.invoke(() -> checkDataAvailable(keys3));
   }
 
   /**
-   * Verify that parallel gateway-sender queue recovers only data enqueued prior to stop command.
+   * Verify that parallel gateway-sender queue recovers only data enqueued prior to stop command
+   * from disk-store.
    *
    * - Region type: PARTITION_PERSISTENT and non-redundant
    * - Gateway sender configured with queue persistence
    *
    * 1. Pause gateway-sender
-   * 2. Run some traffic and verify that data is enqueued in serial gateway-sender queues
+   * 2. Run some traffic and verify that data is enqueued in gateway-sender queues
    * 3. Stop gateway-sender
-   * 4. Restart all servers
-   * 5. Check that queues are not recovered from persistent storage, and that gateway-sender
-   * remained in stopped state
-   * 6. Start gateway-senders
+   * 4. Run some traffic and verify that data is stored in region, and not enqueued
+   * 6. Restart all servers
+   * 7. Check that data that is enqueued prior to stop command is recovered from persistent storage,
+   * and that gateway-sender remained in stopped state
+   * 8. Start gateway-senders
+   * 9. Check that data is not replicated to remote site
    */
   @Test
-  public void testThatStopStateRemainAfterTheRestartOfMembersNotRecover() throws Exception {
-    configureSites(true, "PARTITION_PERSISTENT", "0");
+  @Parameters({"false", "true"})
+  public void testThatStopStateRemainAfterTheRestartAndQueueDataIsRecovered(String isParallel)
+      throws Exception {
+    configureSites(true, "PARTITION_PERSISTENT", "0", isParallel);
 
     executeGfshCommand(CliStrings.PAUSE_GATEWAYSENDER);
     verifyGatewaySenderState(true, true);
@@ -401,11 +391,15 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     server1Site2.invoke(() -> checkQueueSize("ln", keysQueued.size()));
 
     executeGfshCommand(CliStrings.STOP_GATEWAYSENDER);
-    verifyGatewaySenderState(false, false);
+
+    if (isParallel.equals("true")) {
+      verifyGatewaySenderState(false, true);
+    } else {
+      verifyGatewaySenderState(false, false);
+    }
 
     Set<String> keysNotQueued = clientSite2.invoke(() -> doPutsInRange(100, 105));
     clientSite2.invoke(() -> checkDataAvailable(keysNotQueued));
-    server1Site2.invoke(() -> checkQueueSize("ln", 0));
 
     // stop servers on site #2
     server1Site2.stop(false);
@@ -423,16 +417,20 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     thread1.join();
 
     verifyGatewaySenderState(false, false);
-
+    if (isParallel.equals("true")) {
+      server1Site2.invoke(() -> checkQueueSize("ln", keysQueued.size()));
+    }
     executeGfshCommand(CliStrings.START_GATEWAYSENDER);
     verifyGatewaySenderState(true, false);
+
     // Check that data is sent over the gateway-sender
     server1Site2.invoke(() -> checkQueueSize("ln", 0));
     clientSite1.invoke(() -> checkDataAvailable(keysQueued));
     clientSite1.invoke(() -> checkDataNotAvailable(keysNotQueued));
   }
 
-  void configureSites(boolean enableGWSPersistence, String regionShortcut, String redundancy)
+  void configureSites(boolean enableGWSPersistence, String regionShortcut, String redundancy,
+      String isParallel)
       throws Exception {
     String enablePersistenceParameter = "false";
     if (enableGWSPersistence) {
@@ -490,12 +488,12 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     server1Site2 = clusterStartupRule.startServerVM(5, locatorSite2.getPort());
     server2Site2 = clusterStartupRule.startServerVM(6, locatorSite2.getPort());
 
-    // create serial gateway-sender on site #2
+    // create gateway-sender on site #2
     connectGfshToSite(locatorSite2);
     String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
         .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
         .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "1")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "false")
+        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, isParallel)
         .addOption(CliStrings.CREATE_GATEWAYSENDER__ENABLEPERSISTENCE, enablePersistenceParameter)
         .getCommandString();
     gfsh.executeAndAssertThat(command).statusIsSuccess();
@@ -542,13 +540,6 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     server2Site2.invoke(() -> verifySenderState("ln", isRunning, isPaused));
   }
 
-  void verifyGatewaySenderStateOnMember(MemberVM member, boolean isRunning, boolean isPaused) {
-    locatorSite2.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(member.getVM()), "ln", isRunning,
-            isPaused));
-    member.invoke(() -> verifySenderState("ln", isRunning, isPaused));
-  }
-
   private void executeGfshCommand(String cliCommand) {
     String command = new CommandStringBuilder(cliCommand)
         .addOption("id", "ln")
@@ -569,14 +560,14 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
 
   private void checkDataAvailable(Set<String> keys) {
     await()
-        .untilAsserted(() -> assertEquals(keys.size(), isPutAvail(keys)));
+        .untilAsserted(() -> assertEquals(keys.size(), allKeysInRegion(keys)));
   }
 
   private void checkDataNotAvailable(Set<String> keys) {
-    assertNotEquals(keys.size(), isPutAvail(keys));
+    assertNotEquals(keys.size(), allKeysInRegion(keys));
   }
 
-  private int isPutAvail(Set<String> keys) {
+  private int allKeysInRegion(Set<String> keys) {
     Region<String, String> region =
         ClusterStartupRule.clientCacheRule.getCache().getRegion("test1");
     Map<String, String> data = region.getAll(keys);
@@ -598,15 +589,23 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     assertThat(ClusterStartupRule.getCache()).isNotNull();
     InternalCache internalCache = ClusterStartupRule.getCache();
     GatewaySender sender = internalCache.getGatewaySender(senderId);
+    int totalSize = 0;
     Set<RegionQueue> queues = ((AbstractGatewaySender) sender).getQueues();
-    assertFalse(sender.isParallel());
-    int size = 0;
-    if (queues != null) {
-      for (RegionQueue q : queues) {
-        size += q.size();
+    if (sender.isParallel()) {
+      if (queues != null) {
+        for (RegionQueue q : queues) {
+          ConcurrentParallelGatewaySenderQueue prQ = (ConcurrentParallelGatewaySenderQueue) q;
+          totalSize += prQ.size();
+        }
+      }
+    } else {
+      if (queues != null) {
+        for (RegionQueue q : queues) {
+          totalSize += q.size();
+        }
       }
     }
-    assertEquals(numQueueEntries, size);
+    assertEquals(numQueueEntries, totalSize);
   }
 
   /*
@@ -626,13 +625,24 @@ public class SerialGatewaySenderOperationClusterConfigDUnitTest implements Seria
     assertThat(ClusterStartupRule.getCache()).isNotNull();
     InternalCache internalCache = ClusterStartupRule.getCache();
     AbstractGatewaySender sender = (AbstractGatewaySender) internalCache.getGatewaySender(senderId);
-    ConcurrentSerialGatewaySenderEventProcessor abProc =
-        (ConcurrentSerialGatewaySenderEventProcessor) sender.getEventProcessor();
-    List<SerialGatewaySenderEventProcessor> lproc = abProc.getProcessors();
-    assertFalse(lproc.isEmpty());
-    for (SerialGatewaySenderEventProcessor serialProc : lproc) {
-      AbstractGatewaySenderEventProcessor abstProc = serialProc;
-      abstProc.waitForDispatcherToPause();
+    if (sender.isParallel()) {
+      ConcurrentParallelGatewaySenderEventProcessor abProc =
+          (ConcurrentParallelGatewaySenderEventProcessor) sender.getEventProcessor();
+      List<ParallelGatewaySenderEventProcessor> lproc = abProc.getProcessors();
+      assertFalse(lproc.isEmpty());
+      for (ParallelGatewaySenderEventProcessor serialProc : lproc) {
+        AbstractGatewaySenderEventProcessor abstProc = serialProc;
+        abstProc.waitForDispatcherToPause();
+      }
+    } else {
+      ConcurrentSerialGatewaySenderEventProcessor abProc =
+          (ConcurrentSerialGatewaySenderEventProcessor) sender.getEventProcessor();
+      List<SerialGatewaySenderEventProcessor> lproc = abProc.getProcessors();
+      assertFalse(lproc.isEmpty());
+      for (SerialGatewaySenderEventProcessor serialProc : lproc) {
+        AbstractGatewaySenderEventProcessor abstProc = serialProc;
+        abstProc.waitForDispatcherToPause();
+      }
     }
   }
 }
