+++ geode-wan/src/distributedTest/java/org/apache/geode/internal/cache/wan/wancommand/PauseGatewaySenderCommandDUnitTest.java
@@ -214,152 +214,6 @@ public class PauseGatewaySenderCommandDUnitTest implements Serializable {
   }
 
   /**
-   * Test to validate that gateway-sender state is persisted withing cluster configuration
-   * after "pause gateway-sender" command is executed. Additionally, check that resume,
-   * stop and start gateway-sender commands work afterwards as expected.
-   * This behavior is tested in a following way:
-   *
-   * 1. Create gateway-sender on all servers with manual-start set to false (gws will be started
-   * automatically).
-   * 2. Pause gateway-sender using "pause gateway-sender" command.
-   * 3. Stop and then start again all servers and verify that gateway-sender remain paused.
-   * 4. Resume gateway-sender and verify it's state
-   * 5. Stop gateway-sender and verify it's state
-   * 6. Start gateway-sender and verify it's state
-   */
-  @Test
-  public void testPauseGatewaySenderWithClusterConfiguration() {
-    Integer locator1Port = locatorSite1.getPort();
-    // setup servers in Site #1
-    server1 = clusterStartupRule.startServerVM(3, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, locator1Port);
-    server3 = clusterStartupRule.startServerVM(5, locator1Port);
-
-    String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "2")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "false")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__MANUALSTART, "false")
-        .getCommandString();
-    gfsh.executeAndAssertThat(command).statusIsSuccess();
-
-    server1.invoke(() -> startSender("ln"));
-    server2.invoke(() -> startSender("ln"));
-    server3.invoke(() -> startSender("ln"));
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-    server3.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, false));
-
-    command = new CommandStringBuilder(CliStrings.PAUSE_GATEWAYSENDER)
-        .addOption(CliStrings.PAUSE_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK", "OK");
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, true));
-
-    server1.invoke(() -> verifySenderState("ln", true, true));
-    server2.invoke(() -> verifySenderState("ln", true, true));
-    server3.invoke(() -> verifySenderState("ln", true, true));
-
-    server1.stop(true);
-    server2.stop(true);
-
-    server1 = clusterStartupRule.startServerVM(3, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, locator1Port);
-
-    server1.invoke(() -> verifySenderState("ln", true, true));
-    server2.invoke(() -> verifySenderState("ln", true, true));
-    server3.invoke(() -> verifySenderState("ln", true, true));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, true));
-
-    // Stop gateway-sender on all members
-    command = new CommandStringBuilder(CliStrings.RESUME_GATEWAYSENDER)
-        .addOption(CliStrings.RESUME_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK", "OK");
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-    server3.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, false));
-
-    // Stop gateway-sender on all members
-    command = new CommandStringBuilder(CliStrings.STOP_GATEWAYSENDER)
-        .addOption(CliStrings.STOP_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK", "OK");
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-    server2.invoke(() -> verifySenderState("ln", false, false));
-    server3.invoke(() -> verifySenderState("ln", false, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", false, false));
-
-    // Resume gateway-sender on all members
-    command = new CommandStringBuilder(CliStrings.START_GATEWAYSENDER)
-        .addOption(CliStrings.START_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK", "OK");
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-    server3.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, false));
-  }
-
-  /**
    * test to validate that the start gateway sender starts the gateway sender on a group of members
    */
   @Test
@@ -484,103 +338,6 @@ public class PauseGatewaySenderCommandDUnitTest implements Serializable {
     server3.invoke(() -> verifySenderState("ln", true, true));
     server4.invoke(() -> verifySenderState("ln", true, true));
     server5.invoke(() -> verifySenderState("ln", true, false));
-  }
-
-  /**
-   * Test to validate that gateway-sender state is persisted withing cluster configuration
-   * after pause gateway-sender command is executed for particular group.
-   * This behavior is tested in a following way:
-   *
-   * 1. Create gateway-sender on all servers with manual-start set to true (gws won't be started
-   * automatically).
-   * 2. Stop gateway-senders using "pause gateway-sender" command for wanted group of servers.
-   * 3. Stop and then start again servers and verify that gateway-sender in wanted
-   * group of servers remain paused.
-   */
-  @Test
-  public void testPauseGatewaySenderMultipleGroupClusterConfiguration() {
-    int locator1Port = locatorSite1.getPort();
-
-    // setup servers in Site #1
-    server1 = startServerWithGroups(3, "SenderGroup1", locator1Port);
-    server2 = startServerWithGroups(4, "SenderGroup1", locator1Port);
-    server3 = startServerWithGroups(5, "SenderGroup1", locator1Port);
-    server4 = startServerWithGroups(6, "SenderGroup2", locator1Port);
-    server5 = startServerWithGroups(7, "SenderGroup3", locator1Port);
-
-    String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "2")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "false")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__MANUALSTART, "false")
-        .addOption(CliStrings.GROUPS, "SenderGroup1,SenderGroup2,SenderGroup3")
-        .getCommandString();
-    gfsh.executeAndAssertThat(command).statusIsSuccess();
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-    server3.invoke(() -> verifySenderState("ln", true, false));
-    server4.invoke(() -> verifySenderState("ln", true, false));
-    server5.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server4.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server5.getVM()), "ln", true, false));
-
-    command = new CommandStringBuilder(CliStrings.PAUSE_GATEWAYSENDER)
-        .addOption(CliStrings.PAUSE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.GROUP, "SenderGroup1,SenderGroup2")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'SenderGroup1' is updated.")
-        .containsOutput("Cluster configuration for group 'SenderGroup2' is updated.")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK", "OK", "OK");
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server4.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server5.getVM()), "ln", true, false));
-
-    server1.invoke(() -> verifySenderState("ln", true, true));
-    server2.invoke(() -> verifySenderState("ln", true, true));
-    server3.invoke(() -> verifySenderState("ln", true, true));
-    server4.invoke(() -> verifySenderState("ln", true, true));
-    server5.invoke(() -> verifySenderState("ln", true, false));
-
-    server1.stop(false);
-    server3.stop(false);
-    server5.stop(false);
-
-    server1 = startServerWithGroups(3, "SenderGroup1", locator1Port);
-    server3 = startServerWithGroups(5, "SenderGroup1, SenderGroup2", locator1Port);
-    server5 = startServerWithGroups(7, "SenderGroup3", locator1Port);
-
-    server1.invoke(() -> verifySenderState("ln", true, true));
-    server3.invoke(() -> verifySenderState("ln", true, true));
-    server5.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, true));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server5.getVM()), "ln", true, false));
-
   }
 
   private MemberVM startServerWithGroups(int index, String groups, int locPort) {
