+++ geode-wan/src/distributedTest/java/org/apache/geode/internal/cache/wan/wancommand/StartGatewaySenderCommandDUnitTest.java
@@ -139,72 +139,6 @@ public class StartGatewaySenderCommandDUnitTest implements Serializable {
   }
 
   /**
-   * Test to validate that gateway-sender state is persisted withing cluster configuration
-   * after "stop gateway-sender" command is executed.
-   * This behavior is tested in a following way:
-   *
-   * 1. Create gateway-sender on all servers with manual-start set to true (gws will not be started
-   * automatically).
-   * 2. Start gateway-senders using start gateway-sender command.
-   * 3. Stop and then start again all servers and verify that gateway-sender is started.
-   */
-  @Test
-  public void testStartGatewaySenderWithClusterConfiguration() {
-    Integer locator1Port = locatorSite1.getPort();
-
-    // setup servers in Site #1
-    server1 = clusterStartupRule.startServerVM(3, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, locator1Port);
-
-    String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "2")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "false")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__MANUALSTART, "true")
-        .getCommandString();
-    gfsh.executeAndAssertThat(command).statusIsSuccess();
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-    server2.invoke(() -> verifySenderState("ln", false, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", false, false));
-
-    command = new CommandStringBuilder(CliStrings.START_GATEWAYSENDER)
-        .addOption(CliStrings.START_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK");
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-
-    server1.stop(true);
-    server2.stop(true);
-
-    server1 = clusterStartupRule.startServerVM(3, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, locator1Port);
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-  }
-
-  /**
    * Test validates that cluster configuration is not updated when stop gateway-sender is executed
    * per member.
    */
@@ -398,104 +332,6 @@ public class StartGatewaySenderCommandDUnitTest implements Serializable {
     server3.invoke(() -> verifySenderState("ln", true, false));
     server4.invoke(() -> verifySenderState("ln", true, false));
     server5.invoke(() -> verifySenderState("ln", false, false));
-  }
-
-  /**
-   * Test to validate that gateway-sender state is persisted withing cluster configuration
-   * after start gateway-sender command is executed for particular group.
-   * This behavior is tested in a following way:
-   *
-   * 1. Create gateway-sender on all servers with manual-start set to true (gws won't be started
-   * automatically).
-   * 2. Stop gateway-senders using "start gateway-sender" command for wanted group of servers.
-   * 3. Stop and then start again servers and verify that gateway-sender in wanted
-   * group of servers is started.
-   */
-  @Test
-  public void testStartGatewaySenderMultipleGroupClusterConfiguration() {
-    int locator1Port = locatorSite1.getPort();
-
-    // setup servers in Site #1
-    server1 = startServerWithGroups(3, "SenderGroup1", locator1Port);
-    server2 = startServerWithGroups(4, "SenderGroup1", locator1Port);
-    server3 = startServerWithGroups(5, "SenderGroup1, SenderGroup2", locator1Port);
-    server4 = startServerWithGroups(6, "SenderGroup1, SenderGroup2", locator1Port);
-    server5 = startServerWithGroups(7, "SenderGroup3", locator1Port);
-
-    // setup servers in Site #1
-    String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "2")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "true")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__MANUALSTART, "true")
-        .addOption(CliStrings.GROUPS, "SenderGroup1,SenderGroup2,SenderGroup3")
-        .getCommandString();
-    gfsh.executeAndAssertThat(command).statusIsSuccess();
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-    server2.invoke(() -> verifySenderState("ln", false, false));
-    server3.invoke(() -> verifySenderState("ln", false, false));
-    server4.invoke(() -> verifySenderState("ln", false, false));
-    server5.invoke(() -> verifySenderState("ln", false, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server4.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server5.getVM()), "ln", false, false));
-
-    command = new CommandStringBuilder(CliStrings.START_GATEWAYSENDER)
-        .addOption(CliStrings.START_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.GROUP, "SenderGroup1,SenderGroup2")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'SenderGroup1' is updated.")
-        .containsOutput("Cluster configuration for group 'SenderGroup2' is updated.")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK", "OK", "OK");
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server4.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server5.getVM()), "ln", false, false));
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-    server3.invoke(() -> verifySenderState("ln", true, false));
-    server4.invoke(() -> verifySenderState("ln", true, false));
-    server5.invoke(() -> verifySenderState("ln", false, false));
-
-    server1.stop(false);
-    server3.stop(false);
-    server5.stop(false);
-
-    server1 = startServerWithGroups(3, "SenderGroup1", locator1Port);
-    server3 = startServerWithGroups(5, "SenderGroup1, SenderGroup2", locator1Port);
-    server5 = startServerWithGroups(7, "SenderGroup3", locator1Port);
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server3.invoke(() -> verifySenderState("ln", true, false));
-    server5.invoke(() -> verifySenderState("ln", false, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server3.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server5.getVM()), "ln", false, false));
-
   }
 
   /**
