+++ geode-wan/src/distributedTest/java/org/apache/geode/internal/cache/wan/wancommand/StopGatewaySenderCommandDUnitTest.java
@@ -149,141 +149,6 @@ public class StopGatewaySenderCommandDUnitTest implements Serializable {
   }
 
   /**
-   * Test to validate that SERIAL gateway-sender state is persisted withing cluster configuration
-   * after "stop gateway-sender" command is executed.
-   * This behavior is tested in a following way:
-   *
-   * 1. Create gateway-sender on all servers with manual-start set to false (gws will be started
-   * automatically).
-   * 2. Stop gateway-senders using stop gateway-sender command.
-   * 3. Stop and then start again all servers and verify that gateway-sender remain stopped.
-   */
-  @Test
-  public void testStopSerialGatewaySenderWithClusterConfigurationService() {
-    Integer locator1Port = locatorSite1.getPort();
-
-    Properties props = new Properties();
-    server1 = clusterStartupRule.startServerVM(3, props, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, props, locator1Port);
-
-    String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "2")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "false")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__MANUALSTART, "false")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess();
-
-    server1.invoke(() -> startSender("ln"));
-    server2.invoke(() -> startSender("ln"));
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-
-    command = new CommandStringBuilder(CliStrings.STOP_GATEWAYSENDER)
-        .addOption(CliStrings.STOP_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK");
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", false, false));
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-    server2.invoke(() -> verifySenderState("ln", false, false));
-
-    server1.stop(true);
-    server2.stop(true);
-
-    server1 = clusterStartupRule.startServerVM(3, props, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, props, locator1Port);
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-    server1.invoke(() -> verifySenderState("ln", false, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", false, false));
-  }
-
-  /**
-   * Test to validate that PARALLEL gateway-sender state is persisted withing cluster configuration
-   * after "stop gateway-sender" command is executed.
-   * This behavior is tested in a following way:
-   *
-   * 1. Create gateway-sender on all servers with manual-start set to false (gws will be started
-   * automatically).
-   * 2. Stop gateway-senders using stop gateway-sender command.
-   * 3. Stop and then start again all servers and verify that gateway-sender remain stopped.
-   */
-  @Test
-  public void testStopParallelGatewaySenderWithClusterConfigurationService() {
-    Integer locator1Port = locatorSite1.getPort();
-
-    Properties props = new Properties();
-    server1 = clusterStartupRule.startServerVM(3, props, locator1Port);
-    server2 = clusterStartupRule.startServerVM(4, props, locator1Port);
-
-    String command = new CommandStringBuilder(CliStrings.CREATE_GATEWAYSENDER)
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__ID, "ln")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__REMOTEDISTRIBUTEDSYSTEMID, "2")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__PARALLEL, "true")
-        .addOption(CliStrings.CREATE_GATEWAYSENDER__MANUALSTART, "false")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess();
-
-    server1.invoke(() -> startSender("ln"));
-    server2.invoke(() -> startSender("ln"));
-
-    server1.invoke(() -> verifySenderState("ln", true, false));
-    server2.invoke(() -> verifySenderState("ln", true, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", true, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", true, false));
-
-    command = new CommandStringBuilder(CliStrings.STOP_GATEWAYSENDER)
-        .addOption(CliStrings.STOP_GATEWAYSENDER__ID, "ln")
-        .getCommandString();
-
-    gfsh.executeAndAssertThat(command).statusIsSuccess()
-        .containsOutput("Cluster configuration for group 'cluster' is updated")
-        .hasTableSection().hasColumn("Result")
-        .containsExactlyInAnyOrder("OK", "OK");
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server2.getVM()), "ln", false, false));
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-    server2.invoke(() -> verifySenderState("ln", false, false));
-
-    server1.stop(false);
-
-    server1 = clusterStartupRule.startServerVM(3, props, locator1Port);
-
-    server1.invoke(() -> verifySenderState("ln", false, false));
-
-    locatorSite1.invoke(
-        () -> validateGatewaySenderMXBeanProxy(getMember(server1.getVM()), "ln", false, false));
-  }
-
-  /**
    * test to validate that the start gateway sender starts the gateway sender on a member
    */
   @Test
@@ -493,7 +358,6 @@ public class StopGatewaySenderCommandDUnitTest implements Serializable {
     server4.invoke(() -> verifySenderState("ln", false, false));
     server5.invoke(() -> verifySenderState("ln", true, false));
   }
-
 
   @Test
   public void testStopGatewaySender_OneSenderNotRunning() {
